Turn flow complete
The turn flow is now fully functional and works as expected. After the setup phase, the player with the highest roll begins. They must roll the dice and once that's done, they can choose to build, trade, or end their turn. The control is then passed to the next player, and so on and so forth.

- Add EndTurn request used to notify the other players when the current player ended their turn
- Fix round completed check, now takes into account direction of next player