TradeWindow GUI work
Added the left side of the trade window, where the player can build his offer and request, as well a bit of the right side, where other you can see the other players' offers. No functionality is present yet as the GUI work is not done. This is only the GUI window for the trade proposer, the counter-part window displayed to the other players is different and still to come.

- Rename TradeWindow to DomesticTradeWindow
- TradeWindow now represents the window where the player can trade with other players
- Add ResourceWidget, a GUI element that allows the user to select a number of a certain type of resource
- Add TradeOfferItem, a GUI element that shows a counter-offer proposed by another player when trading