Framework for loading saved games & initial Years of Plenty saved game
I started working on the saved games so here's how this works: the games that Jorg wants will be created by the server and associated with 3 dummy players (1, 2, 3). When a player clicks on Browse Games (it should prolly be Load Game but w/e), those games will show up and they can be loaded.
What is left to do is to make sure those games can be only be loaded by "Player 1", "Player 2" and "Player 3", and once a game is loaded, a player must wait until all others are connected.
I started creating the Years of Plenty savegame but so far, it only adds a lot of initial resources to the players.
Bonus: if you load Years of Plenty you will see flag of mother Russia. It is only coincidence. There is no hacking here.

- Fix double session creation by ProgressCardHandler